backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
langchain==0.3.7
langchain-openai==0.2.9
openai==1.54.3
httpx<0.28.0
google-cloud-documentai==2.29.0
chromadb==0.5.18
python-dotenv==1.0.1
